+++ pom.xml
@@ -5,7 +5,7 @@
   <parent>
     <groupId>org.onehippo.cms7</groupId>
     <artifactId>hippo-cms7-release</artifactId>
-    <version>12.6.9</version>
+    <version>14.1.0</version>
   </parent>
 
   <name>XINmods</name>
@@ -63,9 +63,9 @@
 
     <!--***START temporary override of versions*** -->
     <!-- ***END temporary override of versions*** -->
-    <hippo.essentials-components.version>4.6.3</hippo.essentials-components.version>
+    <hippo.essentials-components.version>14.1.0</hippo.essentials-components.version>
+    <essentials.version>14.1.0</essentials.version>
 
-    <essentials.version>4.6.3</essentials.version>
     <jsp-api.version>2.2</jsp-api.version>
     <taglibs.version>1.2.1</taglibs.version>
     <xalan.version>2.7.2</xalan.version>
@@ -90,11 +90,12 @@
 
   <dependencyManagement>
     <dependencies>
+<!--
       <dependency>
         <groupId>org.onehippo.cms7</groupId>
         <artifactId>hippo-essentials-plugin-sdk-api</artifactId>
-        <version>${essentials.version}</version>
       </dependency>
+-->
 
       <!-- Hippo Configuration Management -->
       <dependency>
@@ -344,7 +345,7 @@
         </property>
       </activation>
       <modules>
-        <module>repository-data</module>
+        <!--<module>repository-data</module>-->
         <module>cms</module>
         <module>site</module>
         <module>essentials</module>
