+++ repository-data/config/src/main/resources/hcm-config/hst/configurations.yaml
@@ -1,24 +1,24 @@
 definitions:
   config:
-    /hst:hst/hst:configurations:
-      /common:
-        jcr:primaryType: hst:configuration
-        /hst:sitemap:
-          jcr:primaryType: hst:sitemap
-        /hst:abstractpages:
-          jcr:primaryType: hst:pages
-        /hst:pages:
-          jcr:primaryType: hst:pages
-        /hst:components:
-          jcr:primaryType: hst:components
-        /hst:templates:
-          jcr:primaryType: hst:templates
-        /hst:sitemenus:
-          jcr:primaryType: hst:sitemenus
-        /hst:sitemapitemhandlers:
-          jcr:primaryType: hst:sitemapitemhandlers
-        /hst:catalog:
-          jcr:primaryType: hst:catalog
-      /xinmods:
-        jcr:primaryType: hst:configuration
-        hst:inheritsfrom: [../common]
+      /hst:platform/hst:configurations:
+        /common:
+          jcr:primaryType: hst:configuration
+          /hst:sitemap:
+            jcr:primaryType: hst:sitemap
+          /hst:abstractpages:
+            jcr:primaryType: hst:pages
+          /hst:pages:
+            jcr:primaryType: hst:pages
+          /hst:components:
+            jcr:primaryType: hst:components
+          /hst:templates:
+            jcr:primaryType: hst:templates
+          /hst:sitemenus:
+            jcr:primaryType: hst:sitemenus
+          /hst:sitemapitemhandlers:
+            jcr:primaryType: hst:sitemapitemhandlers
+          /hst:catalog:
+            jcr:primaryType: hst:catalog
+        /xinmods:
+          jcr:primaryType: hst:configuration
+          hst:inheritsfrom: [../common]
