+++ repository-data/config/src/main/resources/hcm-config/hst/configurations/default/templates.yaml
@@ -1,5 +1,5 @@
 definitions:
   config:
-    /hst:hst/hst:configurations/hst:default/hst:templates/essentials-menu:
+    /hst:platform/hst:configurations/hst:default/hst:templates/essentials-menu:
       jcr:primaryType: hst:template
       hst:renderpath: jsp/hstdefault/essentials-menu.jsp
