+++ repository-data/config/src/main/resources/hcm-config/hst/configurations/xinmods/abstractpages.yaml
@@ -1,4 +1,4 @@
 definitions:
   config:
-    /hst:hst/hst:configurations/xinmods/hst:abstractpages:
+    /hst:platform/hst:configurations/xinmods/hst:abstractpages:
       jcr:primaryType: hst:pages
