+++ repository-data/config/src/main/resources/hcm-config/hst/configurations/xinmods/abstractpages/base.yaml
@@ -1,6 +1,6 @@
 definitions:
   config:
-    /hst:hst/hst:configurations/xinmods/hst:abstractpages/base:
+    /hst:platform/hst:configurations/xinmods/hst:abstractpages/base:
       jcr:primaryType: hst:component
       hst:template: base-layout
       /menu:
