+++ repository-data/config/src/main/resources/hcm-config/hst/configurations/xinmods/catalog.yaml
@@ -1,6 +1,6 @@
 definitions:
   config:
-    /hst:hst/hst:configurations/xinmods/hst:catalog:
+    /hst:platform/hst:configurations/xinmods/hst:catalog:
       jcr:primaryType: hst:catalog
       /xinmods-catalog:
         jcr:primaryType: hst:containeritempackage
