+++ repository-data/config/src/main/resources/hcm-config/hst/configurations/xinmods/components.yaml
@@ -1,4 +1,4 @@
 definitions:
   config:
-    /hst:hst/hst:configurations/xinmods/hst:components:
+    /hst:platform/hst:configurations/xinmods/hst:components:
       jcr:primaryType: hst:components
