+++ repository-data/config/src/main/resources/hcm-config/hst/configurations/xinmods/pages.yaml
@@ -1,4 +1,4 @@
 definitions:
   config:
-    /hst:hst/hst:configurations/xinmods/hst:pages:
+    /hst:platform/hst:configurations/xinmods/hst:pages:
       jcr:primaryType: hst:pages
