+++ repository-data/config/src/main/resources/hcm-config/hst/configurations/xinmods/templates.yaml
@@ -1,6 +1,6 @@
 definitions:
   config:
-    /hst:hst/hst:configurations/xinmods/hst:templates:
+    /hst:platform/hst:configurations/xinmods/hst:templates:
       jcr:primaryType: hst:templates
       /base-layout:
         jcr:primaryType: hst:template
