+++ repository-data/config/src/main/resources/hcm-config/hst/configurations/xinmods/workspace/channel.yaml
@@ -1,6 +1,6 @@
 definitions:
   config:
-    /hst:hst/hst:configurations/xinmods/hst:workspace/hst:channel:
+    /hst:platform/hst:configurations/xinmods/hst:workspace/hst:channel:
       jcr:primaryType: hst:channel
       hst:name: XINmods
       hst:type: website
