+++ repository-data/config/src/main/resources/hcm-config/hst/configurations/xinmods/workspace/containers.yaml
@@ -1,6 +1,6 @@
 definitions:
   config:
-    /hst:hst/hst:configurations/xinmods/hst:workspace/hst:containers:
+    /hst:platform/hst:configurations/xinmods/hst:workspace/hst:containers:
       jcr:primaryType: hst:containercomponentfolder
       /base:
         jcr:primaryType: hst:containercomponentfolder
