+++ repository-data/config/src/main/resources/hcm-config/hst/configurations/xinmods/workspace/pages.yaml
@@ -1,5 +1,5 @@
 definitions:
   config:
-    /hst:hst/hst:configurations/xinmods/hst:workspace/hst:pages:
+    /hst:platform/hst:configurations/xinmods/hst:workspace/hst:pages:
       .meta:residual-child-node-category: content
       jcr:primaryType: hst:pages
