+++ repository-data/config/src/main/resources/hcm-config/hst/configurations/xinmods/workspace/sitemap.yaml
@@ -1,5 +1,5 @@
 definitions:
   config:
-    /hst:hst/hst:configurations/xinmods/hst:workspace/hst:sitemap:
+    /hst:platform/hst:configurations/xinmods/hst:workspace/hst:sitemap:
       .meta:residual-child-node-category: content
       jcr:primaryType: hst:sitemap
