+++ repository-data/config/src/main/resources/hcm-config/hst/configurations/xinmods/workspace/sitemenus.yaml
@@ -1,6 +1,6 @@
 definitions:
   config:
-    /hst:hst/hst:configurations/xinmods/hst:workspace/hst:sitemenus:
+    /hst:platform/hst:configurations/xinmods/hst:workspace/hst:sitemenus:
       jcr:primaryType: hst:sitemenus
       /main:
         jcr:primaryType: hst:sitemenu
