+++ repository-data/config/src/main/resources/hcm-config/hst/hosts.yaml
@@ -1,6 +1,6 @@
 definitions:
   config:
-    /hst:hst/hst:hosts:
+    /hst:platform/hst:hosts:
       jcr:primaryType: hst:virtualhosts
       hst:defaultcontextpath: /site
       /dev-localhost:
@@ -15,4 +15,4 @@ definitions:
             .meta:residual-child-node-category: content
             jcr:primaryType: hst:mount
             hst:homepage: root
-            hst:mountpoint: /hst:hst/hst:sites/xinmods
+            hst:mountpoint: /hst:platform/hst:sites/xinmods
