+++ repository-data/config/src/main/resources/hcm-config/hst/sites.yaml
@@ -1,6 +1,6 @@
 definitions:
   config:
-    /hst:hst/hst:sites:
+    /hst:platform/hst:sites:
       .meta:residual-child-node-category: content
       jcr:primaryType: hst:sites
       /xinmods:
