+++ repository-data/config/src/main/resources/hcm-config/main.yaml
@@ -21,18 +21,18 @@ definitions:
         operation: add
         type: string
         value: [editor]
-    /hippo:configuration/hippo:domains/hippofolders/readonly:
-      hipposys:groups:
-        operation: add
-        type: string
-        value: [sitewriters]
-    /hippo:configuration/hippo:domains/preview-documents/readonly:
-      hipposys:groups:
-        operation: add
-        type: string
-        value: [previewusers]
-    /hippo:configuration/hippo:domains/live-documents/readonly:
-      hipposys:groups:
-        operation: add
-        type: string
-        value: [liveusers]
+#    /hippo:configuration/hippo:domains/hippofolders/readonly:
+#      hipposys:groups:
+#        operation: add
+#        type: string
+#        value: [sitewriters]
+#    /hippo:configuration/hippo:domains/preview-documents/readonly:
+#      hipposys:groups:
+#        operation: add
+#        type: string
+#        value: [previewusers]
+#    /hippo:configuration/hippo:domains/live-documents/readonly:
+#      hipposys:groups:
+#        operation: add
+#        type: string
+#        value: [liveusers]
