+++ repository-data/config/src/main/resources/hcm-content/hst/hosts-1.yaml
@@ -1,4 +1,4 @@
-/hst:hst/hst:hosts/dev-localhost/localhost/hst:root/custom-api:
+/hst:platform/hst:hosts/dev-localhost/localhost/hst:root/custom-api:
   jcr:primaryType: hst:mount
   jcr:uuid: 2e3cdb1d-a2ae-4802-9589-18e921f690b2
   hst:alias: custom-api
