+++ repository-data/config/src/main/resources/hcm-content/hst/hosts.yaml
@@ -1,4 +1,4 @@
-/hst:hst/hst:hosts/dev-localhost/localhost/hst:root/api:
+/hst:platform/hst:hosts/dev-localhost/localhost/hst:root/api:
   jcr:primaryType: hst:mount
   jcr:uuid: b42721a2-bfb2-4cc3-be07-8c277f4619f3
   hst:alias: api
